separated components to edit and display a post
