+++ frontend/src/components/project/ProjectProgressPosts/EditProgressPost.js
@@ -1,9 +1,6 @@
-import { Button, Card, IconButton, makeStyles, TextField, Typography } from "@material-ui/core";
+import { Button, Card, makeStyles, TextField } from "@material-ui/core";
 import React, { useState } from "react";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
-import ButtonIcon from "./Buttons/ButtonIcon";
-import { apiRequest } from "../../../public/lib/apiOperations";
-import DateDisplay from "../general/DateDisplay";
+import { apiRequest } from "../../../../public/lib/apiOperations";
 
 const useStyles = makeStyles((theme) => ({
   card: {
@@ -12,47 +9,8 @@ const useStyles = makeStyles((theme) => ({
     paddingRight: theme.spacing(2),
     paddingTop: theme.spacing(1),
   },
-  header: {
-    display: "flex",
-    justifyContent: "space-between",
-  },
-  headerLeft: {
-    display: "flex",
-    flexDirection: "column",
-  },
-  headerRight: {
-    display: "flex",
-    alignItems: "flex-start",
-  },
   textField: {
     marginBottom: theme.spacing(2),
-  },
-  likeButton: {
-    height: 35,
-    padding: theme.spacing(2),
-    margin: theme.spacing(1),
-  },
-  menuButton: {
-    width: 35,
-    height: 35,
-    margin: theme.spacing(1),
-  },
-  heading: {
-    color: theme.palette.primary.main,
-    marginTop: theme.spacing(0.5),
-    marginBottom: theme.spacing(0.5),
-  },
-  lessMargin: {
-    marginTop: theme.spacing(1),
-    marginBottom: theme.spacing(0.5),
-  },
-  border: {
-    borderTop: `1px solid ${theme.palette.grey[500]}`,
-    borderBottom: `1px solid ${theme.palette.grey[500]}`,
-    marginTop: theme.spacing(1),
-    marginBottom: theme.spacing(1),
-    paddingTop: theme.spacing(1),
-    paddingBottom: theme.spacing(1),
   },
   editingButtonsContainer: {
     marginBottom: theme.spacing(2),
@@ -178,63 +136,6 @@ export default function ProgressPost({
           <Button onClick={createPost}>{texts.save}</Button>
           <Button onClick={closeNewPost}>{texts.cancel}</Button>
         </div>
-      </Card>
-    );
-  }
-  //Interface for Viewing
-  else {
-    return (
-      <Card className={classes.card} raised="true">
-        <div className={classes.header}>
-          <div className={classes.headerLeft}>
-            {post.created_at && (
-              <Typography>
-                {<DateDisplay date={new Date(post.created_at)} woTimeAgo />} (
-                {texts.created_lower_case})
-              </Typography>
-            )}
-            {post.updated_at && (
-              <Typography>
-                {<DateDisplay date={new Date(post.updated_at)} woTimeAgo />} ({texts.updated})
-              </Typography>
-            )}
-            {post.event_date && (
-              <Typography>
-                {<DateDisplay date={new Date(post.event_date)} woTimeAgo />} ({texts.event_date})
-              </Typography>
-            )}
-
-            <Typography variant="h5" color="primary">
-              {post.title}
-            </Typography>
-          </div>
-          <div className={classes.headerRight}>
-            {/*Button: Placeholder for LikeButton Component */}
-
-            <Button
-              className={classes.likeButton}
-              variant="contained"
-              color="primary"
-              startIcon={<ButtonIcon icon="like" size={25} color="white" />}
-            >
-              Like • 12
-            </Button>
-
-            <IconButton className={classes.menuButton}>
-              <MoreVertIcon />
-            </IconButton>
-          </div>
-        </div>
-        <TextField
-          className={classes.textField}
-          multiline
-          value={post.content}
-          fullWidth={true}
-          variant="standard"
-          InputProps={{
-            disableUnderline: true,
-          }}
-        />
       </Card>
     );
   }
